Acoplando un solo kernel que es capaz de resolver todos los fiteness para todos los metodos
