+++ gsc/operations.py
@@ -8,8 +8,9 @@ from gsc.kernels.permutationA0001 import PermutationA0001
 from gsc.kernels.crossA0001 import CrossA0001
 from gsc.kernels.mutationA0001 import MutationA0001
 from gsc.kernels.sortA0001 import SortA0001
+from gsc.kernels.fitnessA0001 import FitnessA0001
 
-class Operations(PermutationA0001,CrossA0001,MutationA0001,SortA0001):
+class Operations(PermutationA0001,CrossA0001,MutationA0001,SortA0001,FitnessA0001):
 
     def __init__(self):
         None
@@ -52,6 +53,20 @@ class Operations(PermutationA0001,CrossA0001,MutationA0001,SortA0001):
             return self.exec_permutationA0001()
         else:
             self._population = population
+    
+    def _set_processing_time(self,processing_time):
+        return processing_time
+
+    def _set_machine_sequence(self,machine_sequence):
+        return machine_sequence   
+
+    def _set_due_date(self,due_date):
+        due_date = cp.repeat(cp.expand_dims(cp.array(due_date,dtype=cp.float32),axis=0),self.get_n_samples(),axis=0)
+        return due_date
+    
+    def _set_weights(self,weights):
+        weights = cp.repeat(cp.expand_dims(cp.array(weights,dtype=cp.float32),axis=0),self.get_n_samples(),axis=0) 
+        return weights 
 
     def get_n_samples(self):
         return self._n_samples 
@@ -85,6 +100,18 @@ class Operations(PermutationA0001,CrossA0001,MutationA0001,SortA0001):
 
     def get_population(self): 
         return self._population
+
+    def get_processing_time(self):
+        return self._processing_time
+
+    def get_machine_sequence(self):
+        return self._machine_sequence
+    
+    def get_due_date(self):
+        return self._due_date
+    
+    def get_weights(self):
+        return self._weights
 
     def exec_permutationA0001(self):
         return self._permutationA0001(self.get_n_jobs(),self.get_n_operations(),self.get_n_samples())
@@ -131,4 +158,6 @@ class Operations(PermutationA0001,CrossA0001,MutationA0001,SortA0001):
         y_population,y_sort = self._sortA0001(x_population,x_sort,self.get_n_jobs(),self.get_n_operations(),self.get_n_samples())
         self._set_population(y_population)
         self._set_fitness(y_sort)
-    
+
+    def exec_fitnessA0001(self):
+        fitness = self._fitnessA0001(X,d,w,T,M,digits,n_samples,n_machines)
